Move all funcs and utils into cogs
now the code is much much more easier to read
